Move over NewPageForm logic
