+++ front/app/modules/commercial/customizable_navbar/admin/containers/messages.ts
@@ -29,4 +29,8 @@ export default defineMessages({
     defaultMessage:
       'Are you sure you want to delete this page? This cannot be undone.',
   },
+  addPageButton: {
+    id: 'app.containers.AdminPage.PagesPage.addPageButton',
+    defaultMessage: 'Add page',
+  },
 });
